requirements.txt
numpy
pandas
tensorflow
pillow
tqdm
scipy
pytest
